File Explorer: Move code to reset the icon provider from the Project Explorer
